Trie roughly integrated into TextEditor
I was able to get my Trie working roughly in my Text Editor. I was able to open the keywords.txt, populate the dictionary Trie, and get words to print to the screen. It is very buggy with exceptions when trying to insert the words into the correct location within the editor's text area. I have to create a drop-down/submenu to display the words from the resulting dictionary search.