Added compression functionality
Compression functionality was added to the text editor. It is lossless compression which results in two files being output. one with ***.compressed.txt naming and another with ***.codes.txt naming. These two files serve as the compressed file and the reference file to decode that compressed file. The functionality has progressed through several undocumented stages, but this is the final version.