Integrated sorting code into Text Editor
Code for Insertion sort, Selection sort, Bubble sort, and Quick sort was added to the Text Editor code. This new code is functional. This commit takes care of issues:
#13
#14

Issue #12 needs to be addressed and a new issue #15 is created. The QuickSort code is not working properly after being integrated. Need to debug QuickSort.